Search Page added
Search page is almost 100% finished. It needs a showModal, and the addToLibraryBtns need change of style when newspaper is added and registered to list in Library.

Newspapers are also added to list in Library. Filterbar needs to be activated and an option button for removing from list needs to be added, as well as a showNewspaperModal